--- src/main/java/com/cheering/core/community/constant/BooleanType.java
@@ -1,5 +1,0 @@
-package com.cheering.core.community.constant;
-
-public enum BooleanType {
-    TRUE, FALSE
-}
