--- src/main/java/com/cheering/core/community/constant/Category.java
@@ -1,12 +1,0 @@
-package com.cheering.core.community.constant;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public enum Category {
-    SOCCER("축구");
-
-    private final String korean;
-}
