--- src/main/java/com/cheering/core/community/constant/CommunityType.java
@@ -1,5 +1,0 @@
-package com.cheering.core.community.constant;
-
-public enum CommunityType {
-    PLAYER_COMMUNITY, TEAM_COMMUNITY
-}
