--- src/main/java/com/cheering/core/community/domain/repository/CommunityRepository.java
@@ -1,16 +1,0 @@
-package com.cheering.core.community.domain.repository;
-
-import com.cheering.core.community.domain.Community;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-public interface CommunityRepository extends JpaRepository<Community, Long> {
-
-    @Query("select c from Community c where c.name = :name")
-    List<Community> findByName(@Param("name") String name);
-
-    List<Community> findByNameContainingIgnoreCase(@Param("name") String name);
-
-}
