--- src/main/java/com/cheering/core/community/dto/request/CommunityRequest.java
@@ -1,6 +1,0 @@
-package com.cheering.core.community.dto.request;
-
-public record CommunityRequest(
-        String nickname
-) {
-}
