--- src/main/java/com/cheering/core/community/dto/response/UserCommunityInfoResponse.java
@@ -1,5 +1,0 @@
-package com.cheering.core.community.dto.response;
-
-public record UserCommunityInfoResponse(Long id) {
-
-}
