--- src/main/java/com/cheering/core/user/domain/Role.java
@@ -1,5 +1,0 @@
-package com.cheering.core.user.domain;
-
-public enum Role {
-    ROLE_USER, ROLE_PLAYER, ROLE_ADMIN
-}
