--- src/main/java/com/cheering/core/user/domain/User.java
@@ -1,41 +1,0 @@
-package com.cheering.core.user.domain;
-
-import com.cheering.global.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public class User extends BaseEntity {
-    @Id
-    @GeneratedValue
-    @Column(name = "user_id")
-    private Long id;
-
-    @Column(length = 20)
-    private String name;
-
-    @Column(length = 20)
-    private String nickname;
-
-    @Column(length = 20)
-    private String password;
-
-    @Column(length = 25)
-    private String email;
-
-    @Enumerated(value = EnumType.STRING)
-    private Role role;
-}
