--- src/main/java/com/cheering/core/user/domain/repository/TeamRepository.java
@@ -1,8 +1,0 @@
-package com.cheering.core.user.domain.repository;
-
-import com.cheering.core.user.domain.Team;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<Team, Long> {
-
-}
