--- src/main/java/com/cheering/core/user/domain/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package com.cheering.core.user.domain.repository;
-
-import com.cheering.core.user.domain.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByEmailAndPassword(String email, String password);
-
-    boolean existsUserByEmail(String email);
-
-
-}
