--- src/main/java/com/cheering/core/user/dto/request/SignInRequest.java
@@ -1,7 +1,0 @@
-package com.cheering.core.user.dto.request;
-
-public record SignInRequest(
-        String email,
-        String password
-) {
-}
