--- src/main/java/com/cheering/core/user/dto/response/SignInResponse.java
@@ -1,6 +1,0 @@
-package com.cheering.core.user.dto.response;
-
-public record SignInResponse(
-        Long id
-) {
-}
