--- src/main/java/com/cheering/core/user/dto/response/SignUpResponse.java
@@ -1,4 +1,0 @@
-package com.cheering.core.user.dto.response;
-
-public record SignUpResponse(Long id) {
-}
