+++ src/main/java/com/cheering/domain/community/constant/BooleanType.java
@@ -1,0 +1,5 @@
+package com.cheering.domain.community.constant;
+
+public enum BooleanType {
+    TRUE, FALSE
+}
