+++ src/main/java/com/cheering/domain/community/constant/Category.java
@@ -1,0 +1,12 @@
+package com.cheering.domain.community.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum Category {
+    SOCCER("축구");
+
+    private final String korean;
+}
