+++ src/main/java/com/cheering/domain/community/constant/CommunityType.java
@@ -1,0 +1,5 @@
+package com.cheering.domain.community.constant;
+
+public enum CommunityType {
+    PLAYER_COMMUNITY, TEAM_COMMUNITY
+}
