+++ src/main/java/com/cheering/domain/community/domain/Community.java
@@ -1,0 +1,56 @@
+package com.cheering.domain.community.domain;
+
+import static jakarta.persistence.FetchType.LAZY;
+
+import com.cheering.domain.community.constant.Category;
+import com.cheering.domain.community.constant.CommunityType;
+import com.cheering.domain.community.constant.League;
+import com.cheering.domain.user.domain.Team;
+import com.cheering.global.BaseEntity;
+import com.cheering.domain.user.domain.Player;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+public class Community extends BaseEntity {
+    @Id
+    @GeneratedValue
+    @Column(name = "community_id")
+    private Long id;
+
+    private String name;
+
+    private Long fanCount;
+
+    private String image;
+
+    @Enumerated(value = EnumType.STRING)
+    private Category category;
+
+    @Enumerated(value = EnumType.STRING)
+    private League league;
+
+    @OneToOne(fetch = LAZY, mappedBy = "playerCommunity")
+    private Player player;
+
+    @Enumerated(EnumType.STRING)
+    private CommunityType cType;
+
+    @OneToOne(fetch = LAZY, mappedBy = "teamCommunity")
+    private Team team;
+
+}
