+++ src/main/java/com/cheering/domain/community/dto/request/CommunityRequest.java
@@ -1,0 +1,6 @@
+package com.cheering.domain.community.dto.request;
+
+public record CommunityRequest(
+        String nickname
+) {
+}
