+++ src/main/java/com/cheering/domain/community/dto/response/UserCommunityInfoResponse.java
@@ -1,0 +1,5 @@
+package com.cheering.domain.community.dto.response;
+
+public record UserCommunityInfoResponse(Long id) {
+
+}
