+++ src/main/java/com/cheering/domain/community/service/CommunityService.java
@@ -1,0 +1,303 @@
+package com.cheering.domain.community.service;
+
+import com.cheering.domain.community.constant.BooleanType;
+import com.cheering.domain.community.constant.Category;
+import com.cheering.domain.community.constant.CommunityType;
+import com.cheering.domain.community.constant.League;
+import com.cheering.domain.community.domain.Community;
+import com.cheering.domain.community.domain.UserCommunityInfo;
+import com.cheering.domain.community.domain.repository.CommunityRepository;
+import com.cheering.domain.community.domain.repository.UserCommunityInfoRepository;
+import com.cheering.domain.community.dto.response.CommunityResponse;
+import com.cheering.domain.community.dto.response.FoundCommunitiesResponse;
+import com.cheering.domain.community.dto.response.UserCommunityInfoResponse;
+import com.cheering.domain.user.domain.Team;
+import com.cheering.domain.user.domain.User;
+import com.cheering.domain.user.domain.repository.PlayerRepository;
+import com.cheering.domain.user.domain.repository.UserRepository;
+import com.cheering.global.exception.community.DuplicatedCommunityJoinException;
+import com.cheering.global.exception.community.NotFoundCommunityException;
+import com.cheering.global.exception.constant.ExceptionMessage;
+import com.cheering.global.exception.user.NotFoundUserException;
+import com.cheering.global.util.AwsS3Util;
+import com.cheering.domain.user.domain.Player;
+import com.cheering.domain.user.domain.repository.TeamRepository;
+import java.io.IOException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
+@Service
+@RequiredArgsConstructor
+public class CommunityService {
+
+    private final TeamRepository teamRepository;
+
+    private final UserRepository userRepository;
+    private final CommunityRepository communityRepository;
+    private final UserCommunityInfoRepository userCommunityInfoRepository;
+    private final PlayerRepository playerRepository;
+    private final AwsS3Util awsS3Util;
+
+    public List<FoundCommunitiesResponse> findCommunitiesByName(String name) {
+        List<Community> communities = communityRepository.findByNameContainingIgnoreCase(name);
+
+        List<Community> playerCommunities = communities.stream()
+                .filter(com -> com.getCType().equals(CommunityType.PLAYER_COMMUNITY))
+                .toList();
+
+        User loginUser = getLoginUser();
+        List<UserCommunityInfo> userCommunities = userCommunityInfoRepository.findByUser(loginUser);
+        List<Long> joinCommunityIds = userCommunities.stream().map(com -> com.getCommunity().getId()).toList();
+
+        List<FoundCommunitiesResponse> foundPlayerCommunitiesResponses = generatePlayerCommunityResponse(
+                playerCommunities, joinCommunityIds);
+
+        List<FoundCommunitiesResponse> responseResult = new ArrayList<>(foundPlayerCommunitiesResponses);
+
+        List<Community> teamCommunities = communities.stream()
+                .filter(com -> com.getCType().equals(CommunityType.TEAM_COMMUNITY))
+                .toList();
+
+        if (!teamCommunities.isEmpty()) {
+            List<FoundCommunitiesResponse> foundTeamCommunitiesResponses = generateTeamCommunityResponse(
+                    teamCommunities, joinCommunityIds);
+
+            responseResult.addAll(foundTeamCommunitiesResponses);
+        }
+
+        return responseResult;
+    }
+
+    private List<FoundCommunitiesResponse> generateTeamCommunityResponse(List<Community> teamCommunities,
+                                                                         List<Long> joinCommunityIds) {
+        List<FoundCommunitiesResponse> result = new ArrayList<>();
+        for (Community teamCommunity : teamCommunities) {
+            List<Player> players = teamCommunity.getTeam().getPlayers();
+            List<Community> playerCommunities = getPlayerCommunitiesByPlayers(players);
+
+            List<CommunityResponse> communityResponse = CommunityResponse.ofList(playerCommunities, joinCommunityIds);
+
+            FoundCommunitiesResponse foundCommunitiesResponse;
+
+            if (joinCommunityIds.contains(teamCommunity.getId())) {
+                foundCommunitiesResponse = FoundCommunitiesResponse.of(communityResponse,
+                        teamCommunity, BooleanType.TRUE);
+            } else {
+                foundCommunitiesResponse = FoundCommunitiesResponse.of(communityResponse,
+                        teamCommunity, BooleanType.FALSE);
+            }
+
+            result.add(foundCommunitiesResponse);
+        }
+
+        return result;
+    }
+
+    private List<FoundCommunitiesResponse> generatePlayerCommunityResponse(List<Community> playerCommunities,
+                                                                           List<Long> joinCommunityIds) {
+        List<FoundCommunitiesResponse> result = new ArrayList<>();
+
+        for (Community community : playerCommunities) {
+            if (joinCommunityIds.contains(community.getId())) {
+                CommunityResponse communityResponse = new CommunityResponse(community.getId(), community.getName(),
+                        community.getImage(),
+                        community.getFanCount(),
+                        BooleanType.TRUE);
+
+                FoundCommunitiesResponse foundCommunitiesResponse;
+
+                if (joinCommunityIds.contains(community.getPlayer().getTeam().getTeamCommunity().getId())) {
+                    foundCommunitiesResponse = FoundCommunitiesResponse.of(
+                            List.of(communityResponse), community.getPlayer().getTeam().getTeamCommunity(),
+                            BooleanType.TRUE);
+                } else {
+                    foundCommunitiesResponse = FoundCommunitiesResponse.of(
+                            List.of(communityResponse), community.getPlayer().getTeam().getTeamCommunity(),
+                            BooleanType.FALSE);
+                }
+
+                result.add(foundCommunitiesResponse);
+            } else {
+                CommunityResponse communityResponse = new CommunityResponse(community.getId(), community.getName(),
+                        community.getImage(),
+                        community.getFanCount(),
+                        BooleanType.FALSE);
+
+                FoundCommunitiesResponse foundCommunitiesResponse;
+
+                if (joinCommunityIds.contains(community.getPlayer().getTeam().getTeamCommunity().getId())) {
+                    foundCommunitiesResponse = FoundCommunitiesResponse.of(
+                            List.of(communityResponse), community.getPlayer().getTeam().getTeamCommunity(),
+                            BooleanType.TRUE);
+                } else {
+                    foundCommunitiesResponse = FoundCommunitiesResponse.of(
+                            List.of(communityResponse), community.getPlayer().getTeam().getTeamCommunity(),
+                            BooleanType.FALSE);
+                }
+
+                result.add(foundCommunitiesResponse);
+            }
+        }
+
+        return result;
+    }
+
+    private User getLoginUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String loginId = authentication.getName();
+        return userRepository.findById(Long.valueOf(loginId))
+                .orElseThrow(() -> new NotFoundUserException(ExceptionMessage.NOT_FOUND_USER));
+    }
+
+    @Transactional
+    public UserCommunityInfoResponse joinCommunity(Long communityId, String nickname, MultipartFile file) {
+        User user = getLoginUser();
+
+        Community community = communityRepository.findById(communityId)
+                .orElseThrow(() -> new NotFoundCommunityException(ExceptionMessage.NOT_FOUND_COMMUNITY));
+
+        validateDuplicateJoinCommunity(user, community);
+
+        try {
+            String category = "community/user-community-info-profile";
+            URL url = awsS3Util.uploadFile(file, category);
+            UserCommunityInfo communityUser = UserCommunityInfo.builder()
+                    .nickname(nickname)
+                    .community(community)
+                    .profileImage(url)
+                    .user(user)
+                    .build();
+
+            UserCommunityInfo savedCommunityUser = userCommunityInfoRepository.save(communityUser);
+            return new UserCommunityInfoResponse(savedCommunityUser.getId());
+        } catch (IOException e) {
+            UserCommunityInfo communityUser = UserCommunityInfo.builder()
+                    .nickname(nickname)
+                    .community(community)
+                    .user(user)
+                    .build();
+
+            UserCommunityInfo savedCommunityUser = userCommunityInfoRepository.save(communityUser);
+            return new UserCommunityInfoResponse(savedCommunityUser.getId());
+        }
+    }
+
+    private void validateDuplicateJoinCommunity(User user, Community community) {
+        if (userCommunityInfoRepository.existsByUserAndCommunity(user, community)) {
+            throw new DuplicatedCommunityJoinException(ExceptionMessage.DUPLICATED_JOIN_COMMUNITY);
+        }
+    }
+
+    private static FoundCommunitiesResponse generateFoundCommunitiesResponse(Community community, BooleanType isJoin) {
+        CommunityResponse communityResponse = CommunityResponse.of(community, isJoin);
+        Community teamCommunity = community.getPlayer().getTeam().getTeamCommunity();
+
+        return FoundCommunitiesResponse.of(
+                List.of(communityResponse),
+                teamCommunity,
+                isJoin);
+    }
+
+    private List<Community> getPlayerCommunitiesByPlayers(List<Player> players) {
+        return players.stream().map(Player::getPlayerCommunity).toList();
+    }
+
+    @Transactional
+    public void setData() {
+        String imageUrl = awsS3Util.getPath(
+                "community/user-community-info-profile/0d5211b8-6ee0-4d04-a310-ed1df5dcd89e.png");
+
+        Community psgCommunity = Community.builder()
+                .name("파리 생제르맹")
+                .category(Category.SOCCER)
+                .league(League.FRENCH_LEAGUE1)
+                .image(imageUrl)
+                .cType(CommunityType.TEAM_COMMUNITY)
+                .fanCount(3000L)
+                .build();
+
+        Community tottenhamCommunity = Community.builder()
+                .name("토트넘")
+                .category(Category.SOCCER)
+                .league(League.EPL)
+                .image(imageUrl)
+                .cType(CommunityType.TEAM_COMMUNITY)
+                .fanCount(4000L)
+                .build();
+
+        communityRepository.save(psgCommunity);
+        communityRepository.save(tottenhamCommunity);
+
+        Team teamPSG = Team.builder().players(new ArrayList<>()).teamCommunity(psgCommunity).build();
+        Team teamTottenham = Team.builder().players(new ArrayList<>()).teamCommunity(tottenhamCommunity).build();
+
+        teamRepository.save(teamPSG);
+        teamRepository.save(teamTottenham);
+
+        Community community1 = Community.builder()
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .name("이강인")
+                .fanCount(1L).image(imageUrl).build();
+        Community community2 = Community.builder()
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .name("음바페")
+                .fanCount(2L).image(imageUrl).build();
+        Community community3 = Community.builder()
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .name("아센시오")
+                .fanCount(3L).image(imageUrl).build();
+
+        Community community4 = Community.builder()
+                .name("손흥민")
+                .fanCount(4L)
+                .image(imageUrl)
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .build();
+        Community community5 = Community.builder()
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .name("히샬리송")
+                .fanCount(5L).image(imageUrl).build();
+        Community community6 = Community.builder()
+                .cType(CommunityType.PLAYER_COMMUNITY)
+                .name("메디슨")
+                .fanCount(6L).image(imageUrl).build();
+
+        communityRepository.save(community1);
+        communityRepository.save(community2);
+        communityRepository.save(community3);
+        communityRepository.save(community4);
+        communityRepository.save(community5);
+        communityRepository.save(community6);
+
+        Player playerA1 = Player.builder().playerCommunity(community1).name("이강인").build();
+        Player playerA2 = Player.builder().playerCommunity(community2).name("음바페").build();
+        Player playerA3 = Player.builder().playerCommunity(community3).name("아센시오").build();
+
+        playerA1.connectTeam(teamPSG);
+        playerA2.connectTeam(teamPSG);
+        playerA3.connectTeam(teamPSG);
+
+        playerRepository.save(playerA1);
+        playerRepository.save(playerA2);
+        playerRepository.save(playerA3);
+
+        Player playerB1 = Player.builder().playerCommunity(community4).name("손흥민").build();
+        Player playerB2 = Player.builder().playerCommunity(community5).name("히샬리송").build();
+        Player playerB3 = Player.builder().playerCommunity(community6).name("메디슨").build();
+
+        playerB1.connectTeam(teamTottenham);
+        playerB2.connectTeam(teamTottenham);
+        playerB3.connectTeam(teamTottenham);
+
+        playerRepository.save(playerB1);
+        playerRepository.save(playerB2);
+        playerRepository.save(playerB3);
+    }
+}
