+++ src/main/java/com/cheering/domain/user/domain/Role.java
@@ -1,0 +1,5 @@
+package com.cheering.domain.user.domain;
+
+public enum Role {
+    ROLE_USER, ROLE_PLAYER, ROLE_ADMIN
+}
