+++ src/main/java/com/cheering/domain/user/domain/repository/PlayerRepository.java
@@ -1,0 +1,7 @@
+package com.cheering.domain.user.domain.repository;
+
+import com.cheering.domain.user.domain.Player;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlayerRepository extends JpaRepository<Player, Long> {
+}
