+++ src/main/java/com/cheering/domain/user/domain/repository/TeamRepository.java
@@ -1,0 +1,8 @@
+package com.cheering.domain.user.domain.repository;
+
+import com.cheering.domain.user.domain.Team;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamRepository extends JpaRepository<Team, Long> {
+
+}
