+++ src/main/java/com/cheering/domain/user/domain/repository/UserRepository.java
@@ -1,0 +1,17 @@
+package com.cheering.domain.user.domain.repository;
+
+import com.cheering.domain.user.domain.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+
+    Optional<User> findByEmailAndPassword(String email, String password);
+
+    boolean existsUserByEmail(String email);
+
+
+}
