+++ src/main/java/com/cheering/domain/user/dto/request/SignInRequest.java
@@ -1,0 +1,7 @@
+package com.cheering.domain.user.dto.request;
+
+public record SignInRequest(
+        String email,
+        String password
+) {
+}
