+++ src/main/java/com/cheering/domain/user/dto/response/SignInResponse.java
@@ -1,0 +1,6 @@
+package com.cheering.domain.user.dto.response;
+
+public record SignInResponse(
+        Long id
+) {
+}
