+++ src/main/java/com/cheering/domain/user/dto/response/SignUpResponse.java
@@ -1,0 +1,4 @@
+package com.cheering.domain.user.dto.response;
+
+public record SignUpResponse(Long id) {
+}
