+++ src/test/java/com/cheering/user/service/UserServiceTest.java
@@ -4,13 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.cheering.core.user.service.UserService;
+import com.cheering.domain.user.service.UserService;
 import com.cheering.global.exception.user.DuplicatedEmailException;
 import com.cheering.global.exception.user.NotFoundUserException;
-import com.cheering.core.user.domain.User;
-import com.cheering.core.user.domain.repository.UserRepository;
-import com.cheering.core.user.dto.request.SignInRequest;
-import com.cheering.core.user.dto.request.SignUpRequest;
+import com.cheering.domain.user.domain.User;
+import com.cheering.domain.user.domain.repository.UserRepository;
+import com.cheering.domain.user.dto.request.SignInRequest;
+import com.cheering.domain.user.dto.request.SignUpRequest;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
